refactor(app): isolate library manager state
- Move catalog, scope, loading, and update orchestration into a composable
- Ignore stale asynchronous refresh results
- Report catalog failures through the shared toast system
